+++ storage_stats/__main__.py
@@ -25,8 +25,14 @@ def main(args=None):
 
     # Process CLI arguments #
     ap = argparse.ArgumentParser(prog="storage_stats",
-                                 description="Calculates file size statistics for the specified folder")
+                                 description="Calculates file size statistics for the specified folder.",
+                                 epilog="MAPFILE should be a text file with one group of similar file"
+                                        " extensions per line, separated by commas. Each line should be in"
+                                        " lowercase and take the form: \".main_ext,.alt1,.alt2,etc\". For example: "
+                                        "\".tiff,.tif\"")
     ap.add_argument("path", help="the folder to characterise")
+    ap.add_argument("-e", dest="mapfile", default=None,
+                    help="user file overriding similar extension mappings")
     ap.add_argument("-o", dest="output", help="CSV file to output statistics too")
     ap.add_argument("--no-recursion", dest="recursive", action="store_false",
                     help="do not include sub-folders in stats")
@@ -39,7 +45,7 @@ def main(args=None):
 
     if args.path:
         # process the specified directory and print the stats
-        characteriser = ss.Characteriser()
+        characteriser = ss.Characteriser(args.mapfile)
         characteriser.process_directory(args.path, args.recursive, args.timing)
 
         if not args.silent: